Exam Percentage Change Seconds to Minutes

+++ AndersonExamData/DExaminee.cs
@@ -13,7 +13,7 @@ namespace AndersonExamData
         #endregion
 
         #region READ
-        public decimal Percentage(int examineeId)
+        public int Percentage(int examineeId)
         {
             using (var context = new Context())
             {
@@ -29,7 +29,7 @@ namespace AndersonExamData
                 if (takenExamSummary == null || takenExamSummary.Count() == 0)
                     return 0;
                 var totalPercentage = takenExamSummary?.Average(a => ((a.Score * 100) / a.Total)) ?? 0;
-                return (decimal)totalPercentage;
+                return (int)totalPercentage;
             }
         }
         #endregion
